requirements.txt
streamlit
openai
faiss-cpu
numpy
pandas
python-dotenv
streamlit-authenticator
